requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
openai==1.59.7
faiss-cpu==1.9.0.post1
numpy==2.2.1
pypdf==5.1.0
python-docx==1.1.2
streamlit==1.41.1
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
reportlab==4.2.5
python-dotenv==1.0.1
